payment

+++ internal/db/query/cart.sql
@@ -25,8 +25,7 @@ SELECT
     p.discount,
 
     -- قیمت نهایی بعد از تخفیف
-    ROUND(p.price * (1 - COALESCE(p.discount, 0) / 100)) AS final_price,
-
+    ROUND(p.price * (1 - COALESCE(p.discount, 0) / 100.0))::bigint AS final_price,
     pi.url AS image_url
 
 FROM cart_items ci
